Fix more issues w/ movielist naming convention
