+++ app/components/movieList/MovieList.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import './movieList-style';
+
+const MovieList = (props) => {
+
+  return (
+    <div>
+      I'm a movielist div
+    </div>
+  )
+}
+
+export default MovieList;
